requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
flask-socketio==5.3.4
python-socketio==5.9.0
simple-websocket==1.0.0
requests==2.31.0

